🏗️ Architect: Modularized astronomical math utilities
Extracted duplicated geometric altitude, azimuth, and separation logic from `DiscoveryService` and `visibility.py` into a new centralized utility module `apts/utils/astronomy.py`.

Key changes:
- Created `apts/utils/astronomy.py` with vectorized NumPy implementations of `vectorized_geometric_altaz` and `vectorized_angular_separation`.
- Moved `calculate_refraction` and other geometric compute functions to `apts/utils/astronomy.py`.
- Refactored `DiscoveryService._populate_bulk_data` in `apts/discovery/service.py` by extracting sub-methods for better readability and maintainability.
- Updated `apts/objects/visibility.py` to use the shared `vectorized_geometric_altaz` utility.
- Maintained backward compatibility in `apts/objects/utils.py` via re-exports.

This refactor adheres to the Single Responsibility Principle and improves code reuse across the repository.
